+++ AmazonChimeSDK/AmazonChimeSDK/internal/utils/DeviceUtils.swift
@@ -44,6 +44,7 @@ import UIKit
         appInfo.platform_version = UnsafePointer<Int8>((info.platformVersion as NSString).utf8String)
         appInfo.client_source = UnsafePointer<Int8>((info.clientSource as NSString).utf8String)
         appInfo.chime_sdk_version = UnsafePointer<Int8>((info.chimeSdkVersion as NSString).utf8String)
+        appInfo.client_utc_offset = UnsafePointer<Int8>((info.clientUtcOffset as NSString).utf8String)
         return appInfo
     }
 
@@ -60,7 +61,16 @@ import UIKit
         appInfo.deviceMake = "apple"
         appInfo.clientSource = "amazon-chime-sdk"
         appInfo.chimeSdkVersion = Versioning.sdkVersion()
+        appInfo.clientUtcOffset = getClientUtcOffset()
         return appInfo
+    }
+
+    static private func getClientUtcOffset() -> String {
+        let seconds = TimeZone.current.secondsFromGMT()
+        let hours = seconds/3600
+        let minutes = abs(seconds/60) % 60
+        let tz = String(format: "%+.2d:%.2d", hours, minutes)
+        return tz
     }
 
     private static func mediaLibInfo() -> String {