+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+### Added
+* Pass client UTC offset to audio and video client for metrics.
+
+### Changed
+
+### Removed
+
+### Fixed
+
 ## [0.23.0] - 2023-03-16
 
 ### Added